Tweaks namespace for viewmodels and completes layout of add who to group form
