+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/BulkUpload/AddWhoToGroup.cshtml
@@ -1,0 +1,56 @@
+﻿@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.BulkUpload
+@model AddWhoToGroupViewModel
+@{
+    var errorHasOccurred = !ViewData.ModelState.IsValid;
+    ViewData["Title"] = (errorHasOccurred? "Error: " : "")  + "Add which delegates to a group";
+}
+<div class="nhsuk-grid-row">
+    <div class="nhsuk-grid-column-full">
+        @if (errorHasOccurred)
+        {
+            <vc:error-summary order-of-property-names="@(new []{ nameof(AddToGroupViewModel.ExistingGroupId), nameof(AddToGroupViewModel.NewGroupName) })" />
+        }
+        <form class="nhsuk-u-margin-bottom-3" method="post" asp-action="SubmitAddWhoToGroup" enctype="multipart/form-data">
+            <fieldset class="nhsuk-fieldset">
+                <legend class="nhsuk-fieldset__legend nhsuk-fieldset__legend--l">
+                    <h1 class="nhsuk-fieldset__heading">
+                        Who should be added to this group?
+                    </h1>
+                </legend>
+                <div class="nhsuk-hint nhsuk-u-margin-bottom-2">
+                    Choose who you want to be added to the group @Model.GroupName
+                </div>
+                <div class="nhsuk-hint">
+                    Select one option
+                </div>
+                <div class="nhsuk-form-group">
+                    <div class="nhsuk-radios">
+                        <div class="nhsuk-radios__item">
+                            <input class="nhsuk-radios__input" id="option-1" asp-for="IncludeUpdatedDelegates" type="radio" value="true" aria-describedby="option-1-hint">
+                            <label class="nhsuk-label nhsuk-radios__label" for="option-1">
+                                Add all delegates in my uploaded sheet to this group
+                            </label>
+                            <div class="nhsuk-hint nhsuk-radios__hint" id="option-1-hint">
+                                @Model.ToProcessCount delegates will be added to the group @Model.GroupName
+                            </div>
+                        </div>
+
+                        <div class="nhsuk-radios__item">
+                            <input class="nhsuk-radios__input" id="option-2" asp-for="IncludeUpdatedDelegates" type="radio" value="false" aria-describedby="option-1-hint">
+                            <label class="nhsuk-label nhsuk-radios__label" for="option-2">
+                                Only add newly registered delegates to this group 
+                            </label>
+                            <div class="nhsuk-hint nhsuk-radios__hint" id="option-2-hint">
+                                @Model.ToRegisterCount delegates will be added to the group @Model.GroupName
+                            </div>
+                        </div>
+                    </div>
+                </div>
+
+            </fieldset>
+            <button class="nhsuk-button" type="submit">Next</button>
+        </form>
+
+        <vc:cancel-link asp-controller="BulkUpload" asp-action="CancelUpload" />
+    </div>
+</div>
