+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/BulkUpload/UploadCompleted.cshtml
@@ -1,4 +1,4 @@
-﻿@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates
+﻿@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.BulkUpload
 @model BulkUploadPreProcessViewModel
 
 @{
